working on 11

+++ src/utils/Api.js
@@ -47,22 +47,40 @@ class Api {
       .then(res => res.ok ? res.json() : Promise.reject('Error' + res.statusText))
   }
 
+  changeLikeCardStatus(cardId, like) {
+    if(like) {
+      return fetch(`${this._baseUrl}/cards/likes/${cardId}`, {
+        headers: this._headers,
+        method: "PUT",
+        body: JSON.stringify({
+          like
+        })
+      })
+    } else {
+      return fetch(`${this._baseUrl}/cards/likes/${cardId}`, {
+        headers: this._headers,
+        method: "DELETE"
+      })
+      .then(res => res.ok ? res.json() : Promise.reject('Error' + res.statusText))
+    }
+  }
+
   //PUT https://around.nomoreparties.co/v1/group-4/cards/likes/cardId
-  cardLike(cardId) {
-    return fetch(`${this._baseUrl}/cards/likes/${cardId}`, {
-      headers: this._headers,
-      method: "PUT"
-    })
-      .then(res => res.ok ? res.json() : Promise.reject('Error' + res.statusText))
-  }
+  // cardLike(cardId) {
+  //   return fetch(`${this._baseUrl}/cards/likes/${cardId}`, {
+  //     headers: this._headers,
+  //     method: "PUT"
+  //   })
+  //     .then(res => res.ok ? res.json() : Promise.reject('Error' + res.statusText))
+  // }
   //DELETE https://around.nomoreparties.co/v1/group-4/cards/likes/cardId
-  cardUnlike(cardId) {
-    return fetch(`${this._baseUrl}/cards/likes/${cardId}`, {
-      headers: this._headers,
-      method: "DELETE"
-    })
-      .then(res => res.ok ? res.json() : Promise.reject('Error' + res.statusText))
-  }
+  // cardUnlike(cardId) {
+  //   return fetch(`${this._baseUrl}/cards/likes/${cardId}`, {
+  //     headers: this._headers,
+  //     method: "DELETE"
+  //   })
+  //     .then(res => res.ok ? res.json() : Promise.reject('Error' + res.statusText))
+  // }
 
 
   //PATCH https://around.nomoreparties.co/v1/group-4/users/me
@@ -80,7 +98,7 @@ class Api {
   }
 
   //PATCH https://around.nomoreparties.co/v1/groupId/users/me/avatar
-  setUserAvatar({ avatar }) {    
+  setUserAvatar({ avatar }) {
     return fetch(`${this._baseUrl}/users/me/avatar`, {
       headers: this._headers,
       method: "PATCH",
